feat: support sub-attributes after value paths.
Allow comparison operations on sub-attributes following value paths.
e.g. `emails[type eq "work"].value eq "user@example.com"`

The tokenizer regex was updated to allow for a trailing dot (.) after a closing square bracket (]) when searching for "Bracket" tokens.
`readValFilter` was then updated to look for any dot after the closing bracket and a following "Word" token. When this happens an implicit `and` op is returned.

One issue with this approach is that any `and` ops before or after this syntax will result in nested `and` ops instead of a single op with more than two filters. Check the last test added to `parse.test.ts` for an example. However, I couldn't figure out a neat way to flattening that use case without a much more disruptive changeset.

fixes: #96