Add expand utility function
This utility applies the distributive property of boolean `and`s over
`or`s on a given AST, thereby expanding out bracketed expressions.

This also makes the `LogExp` type a union of the new `OrExp` and
`AndExp` types.

Finally, this adds to the README an example of the existing `flatten`
function, as well as for the new `expand` function.